requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
google-genai
langchain-chroma
langchain-community
python-dotenv
